+++ app.js
@@ -93,13 +93,14 @@ swaggerTools.initializeMiddleware(swaggerConfig, function(middleware) {
   const options = {
     user: dbUsername,
     pass: dbPassword,
+    authSource: 'admin',
     maxPoolSize: 10, // Maintain up to 10 socket connections
     connectTimeoutMS: 10000, // Give up initial connection after 10 seconds
     socketTimeoutMS: 45000, // Close sockets after 45 seconds of inactivity
   };
   defaultLog.info('Connecting to:', dbConnection);
   mongoose.Promise = global.Promise;
-  (mongoose.connect(dbConnection, options).then(() => {
+  (mongoose.connect(dbConnection, options).then(async () => {
     defaultLog.info('Database connected');
 
     // Load database models
@@ -126,6 +127,8 @@ swaggerTools.initializeMiddleware(swaggerConfig, function(middleware) {
     require('./api/helpers/models/topic');
     require('./api/helpers/models/emailSubscribe');
     defaultLog.info('db model loading done.');
+
+    await mongoose.connection.syncIndexes();
 
     // Build text index.
     databaseIndexes.generateTextIndex();